dil/pyproject.toml
[tool.poetry]
name = "dil"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
tranco = "^0.6"
pandas = "^1.4.1"
psycopg2 = "^2.9.3"
tqdm = "^4.63.0"
h2o = "^3.36.0"
Django = "^4.0.3"
Pillow = "^9.0.1"
diff-match-patch = "^20200713"
python-Levenshtein = "^0.12.2"
matplotlib-venn = "^0.11.7"
scipy = "1.8"
pyzmq = "^23.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
